requirements.txt
schedule
python-dotenv
